Bitmaps and Autotile
Added an autotile system for the levels. The level terrain will be generated according to the obstacles on it; each tile has a corresponding 3x3 bitmap in a separate image file that dictates when this tile is used.

Each Tileset needs to have a corresponding bitmap image file in order for this system to work properly.